fixed intro more pages
