+++ mobile-app/app/src/main/java/com/semirsuljevic/raiffaisenmobileapp/ui/screens/intro_navbar/intro_more/OnBoarding.kt
@@ -5,6 +5,7 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
 import androidx.compose.material.TextButton
 import androidx.compose.runtime.Composable
@@ -25,6 +26,7 @@ import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.PagerState
 import com.google.accompanist.pager.rememberPagerState
 import com.semirsuljevic.raiffaisenmobileapp.R
+import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Black
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Gray400
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.White
 import com.semirsuljevic.raiffaisenmobileapp.ui.theme.Yellow400
@@ -61,50 +63,54 @@ fun OnBoardingTabs(
             }
         }
     }
-    Column (
-        modifier = Modifier.fillMaxHeight(),
-        horizontalAlignment = Alignment.CenterHorizontally
+    Scaffold (
+        backgroundColor = Black
     ){
-        HorizontalPager(
-            state = pagerState,
-            dragEnabled = true
-        ) {
-                page ->
-            when(page) {
-                0 -> OnBoardingTabItem(
+        Column (
+            modifier = Modifier.fillMaxHeight(),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ){
+            HorizontalPager(
+                state = pagerState,
+                dragEnabled = true
+            ) {
+                    page ->
+                when(page) {
+                    0 -> OnBoardingTabItem(
                         title = stringResource(id = R.string.onboarding_item1_title),
                         description = stringResource(id = R.string.onboarding_item1_desc),
                         navController = navController,
                         image = R.drawable.onboarding_item1
                     )
-                1 -> OnBoardingTabItem(
+                    1 -> OnBoardingTabItem(
                         title = stringResource(id = R.string.onboarding_item2_title),
                         description = stringResource(id = R.string.onboarding_item2_desc),
                         navController = navController,
-                    image = R.drawable.onboarding_item2
+                        image = R.drawable.onboarding_item2
                     )
-                2 -> OnBoardingTabItem(
+                    2 -> OnBoardingTabItem(
                         title = stringResource(id = R.string.onboarding_item3_title),
                         description = stringResource(id = R.string.onboarding_item3_desc),
                         navController = navController,
                         image = R.drawable.onboarding_item3
                     )
-                3 -> OnBoardingTabItem(
+                    3 -> OnBoardingTabItem(
                         title = stringResource(id = R.string.onboarding_item4_title),
                         description = stringResource(id = R.string.onboarding_item4_desc),
                         navController = navController,
                         image = R.drawable.onboarding_item4
                     )
-                4 -> OnBoardingTabItem(
+                    4 -> OnBoardingTabItem(
                         title = stringResource(id = R.string.onboarding_item5_title),
                         description = stringResource(id = R.string.onboarding_item5_desc),
                         navController = navController,
                         image = R.drawable.onboarding_item5
                     )
+                }
             }
+            Spacer(modifier = Modifier.height(12.dp))
+            IndexList(viewModel = viewModel)
         }
-        Spacer(modifier = Modifier.height(12.dp))
-        IndexList(viewModel = viewModel)
     }
 }
